+++ tests/Unit/Application/Controller/ViewTest.php
@@ -503,36 +503,6 @@ class ViewTest extends \OxidTestCase
         $this->assertEquals($isRCVersion, $oView->isRCVersion());
     }
 
-    /**
-     * testShowBetaBote data provider.
-     */
-    public function _dptestShowBetaNote()
-    {
-        return array(
-            array(false, false, false),
-            array(true, false, true),
-            array(false, true, true),
-            array(true, true, true),
-        );
-    }
-
-    /**
-     * @dataProvider _dptestShowBetaNote
-     */
-    public function testShowBetaNote($isBetaVersion, $isRCVersion, $showBetaNote)
-    {
-        $oxView = oxNew('oxView');
-        if (!$oxView->showBetaNote()) {
-            $this->markTestSkipped('there is no real beta note for this version');
-        }
-
-        $oView = $this->getMock(\OxidEsales\Eshop\Core\Controller\BaseController::class, array('isBetaVersion', 'isRCVersion'), array(), '', false);
-        $oView->expects($this->any())->method('isBetaVersion')->will($this->returnValue($isBetaVersion));
-        $oView->expects($this->any())->method('isRCVersion')->will($this->returnValue($isRCVersion));
-
-        $this->assertEquals($showBetaNote, $oView->showBetaNote());
-    }
-
     public function testEditionIsNotEmpty()
     {
         //edition is always set